Extract fs init code
